+++ NEWS.md
@@ -13,6 +13,7 @@
 * Fixed issue in `get_data()` for *clmm2*-models.
 * Fixed issue in `get_data()` for models that used the `lspline()`-function.
 * Fixed issue in `get_statistic()` for *multinom* models.
+* Fixed issue in `get_priors()` for *stanreg*  models with flat intercept priors.
 
 # insight 0.8.0